Style: 리팩터링
