+++ 류미성/moonbucks-menu/src/js/index.js
@@ -22,16 +22,8 @@
 // - [] 품절 해당 메뉴의 상태값이 페이지에 그려진다.
 //  -[] 클릭이벤트에서 가장 가까운 li태그의 class속성 값에 sold-out을 추가한다.
 
-const $ = (selector) => document.querySelector(selector);
-
-const store = {
-	setLocalStorage(menu) {
-		localStorage.setItem("menu", JSON.stringify(menu));
-	},
-	getLocalStorage() {
-		return JSON.parse(localStorage.getItem("menu"));
-	},
-};
+import { $ } from "./utils/dom.js";
+import store from "./store/index.js";
 
 function App() {
 	this.menu = {
@@ -48,6 +40,7 @@ function App() {
 			this.menu = store.getLocalStorage();
 		}
 		render();
+		initEventListeners();
 	};
 
 	const render = () => {
@@ -81,7 +74,7 @@ function App() {
 	};
 
 	const updateMenuCount = () => {
-		const menuCount = $("#menu-list").querySelectorAll("li").length;
+		const menuCount = this.menu[this.currentCategory].length;
 		$(".menu-count").innerText = `총 ${menuCount} 개`;
 	};
 
@@ -103,7 +96,7 @@ function App() {
 		const updatedMenuName = prompt("메뉴명을 수정하세요", $menuName.innerText); // 인자 사용 유의하기
 		this.menu[this.currentCategory][menuId].name = updatedMenuName;
 		store.setLocalStorage(this.menu);
-		$menuName.innerText = updatedMenuName; //e를 변수로 활용
+		render();
 	};
 
 	const removeMenuName = (e) => {
@@ -111,8 +104,7 @@ function App() {
 			const menuId = e.target.closest("li").dataset.menuId;
 			this.menu[this.currentCategory].splice(menuId, 1);
 			store.setLocalStorage(this.menu);
-			e.target.closest("li").remove();
-			updateMenuCount();
+			render();
 		}
 	};
 
@@ -124,48 +116,49 @@ function App() {
 		render();
 	};
 
-	$("#menu-list").addEventListener("click", (e) => {
-		if (e.target.classList.contains("menu-edit-button")) {
-			updateMenuName(e);
-			return;
-		}
+	const initEventListeners = () => {
+		$("#menu-list").addEventListener("click", (e) => {
+			if (e.target.classList.contains("menu-edit-button")) {
+				updateMenuName(e);
+				return;
+			}
 
-		if (e.target.classList.contains("menu-remove-button")) {
-			removeMenuName(e);
-			return;
-		}
+			if (e.target.classList.contains("menu-remove-button")) {
+				removeMenuName(e);
+				return;
+			}
 
-		if (e.target.classList.contains("menu-sold-out-button")) {
-			soldOutMenu(e);
-			return;
-		}
-	});
+			if (e.target.classList.contains("menu-sold-out-button")) {
+				soldOutMenu(e);
+				return;
+			}
+		});
 
-	//form태그가 자동으로 전송되는걸 막아준다.
-	$("#menu-form").addEventListener("submit", (e) => {
-		e.preventDefault();
-	});
+		$("#menu-form").addEventListener("submit", (e) => {
+			e.preventDefault();
+		});
 
-	$("#menu-submit-button").addEventListener("click", addMenuName);
+		$("#menu-submit-button").addEventListener("click", addMenuName);
 
-	// 메뉴의 이름을 입력받는건
-	$("#menu-name").addEventListener("keypress", (e) => {
-		if (e.key !== "Enter") {
-			// 처음에 엔터를 눌러도 alert 안뜨게
-			return;
-		}
-		addMenuName();
-	});
+		// 메뉴의 이름을 입력받는건
+		$("#menu-name").addEventListener("keypress", (e) => {
+			if (e.key !== "Enter") {
+				// 처음에 엔터를 눌러도 alert 안뜨게
+				return;
+			}
+			addMenuName();
+		});
 
-	$("nav").addEventListener("click", (e) => {
-		const isCategoryButton = e.target.classList.contains("cafe-category-name");
-		if (isCategoryButton) {
-			const categoryName = e.target.dataset.categoryName;
-			this.currentCategory = categoryName;
-			$("#category-title").innerText = `${e.target.innerText} 메뉴관리`;
-			render();
-		}
-	});
+		$("nav").addEventListener("click", (e) => {
+			const isCategoryButton = e.target.classList.contains("cafe-category-name");
+			if (isCategoryButton) {
+				const categoryName = e.target.dataset.categoryName;
+				this.currentCategory = categoryName;
+				$("#category-title").innerText = `${e.target.innerText} 메뉴관리`;
+				render();
+			}
+		});
+	};
 }
 const app = new App();
 app.init();
