+++ 류미성/moonbucks-menu/src/js/utils/dom.js
@@ -1,0 +1,1 @@
+export const $ = (selector) => document.querySelector(selector);
